Implement most feedback of Kulla
